seperate autoscaler business logic from httpClient logic
